Move Gcc_backend declaration to header file
gcc/rust/ChangeLog:

	* rust-backend.h: Include "rust-gcc.h".
	(class Gcc_backend): Move to ...
	* rust-gcc.cc (class Gcc_backend): ... here.
